+++ resqpy/property/stringlookup.py
