+++ resqpy/property/wellintervalproperty.py
