+++ resqpy/property/welllogcollection.py
